rename StopBasedDrtRoutingModule to DrtRoutingModule
